Added Player Movement
Created a car that the player can control.
This ended up being much much much much harder than I anticipated. I could
not get the rotations to work how I wanted so I switched to a different
approach. I think what I have now is ok, but it still feels terrible.

Player uses a and d to steer and w and s to drive forwards and back. The
car should have acceleration so it takes a moment for the car to get to
top speed. Same with turning, need to turn for a second to get the max
turn angle. All these inputs should go back to 0 when the button is not
being pushed (the feel on this is still a bit off though).

Also added a mage that can fire spells for the player to dodge. Made 2
attack types. Pulse should fire out in 8 directions at once. Spiral should
fire out one in one direction, then the next direction, etc.