Created More Map Pieces
Made some angled hallways and some rooms.
So far, all enter from the left and exit to the right.
Each new room adds a lot of new code (same code copy-pasted and change
some edge case). This makes it very error prone, hard to read, hard to
debug.

A different setup would be to manually make a kind of mask.
For example, a 20x20 array of numbers, 0 is floor, 1 is wall, 2 is empty
space, 3 is entrance, 4 is exit. Then I would edit these by hand in some
kind of editor and build a bunch of these objects in a folder instead of
in code.
Could be a sprite and just drop into the correct array.
The arrays would be leftRight for entrance on left, exit on right or
leftDown for entrance on left, exit down.