Created More Map Pieces
In Mage:
Added an attack type so they don't attack at all
When mages are hit by the player, they go flying
When they hit the ground again, they should explode or something (but they
don't yet)

In MapGen:
Tried making more custom pieces for the race track. They were quite time
consuming, inaccurate, and took a lot of memory (72kb for a 10x10 tile vs
1kb for the script). I don't know if that would be a problem for a web
build or not.

Instead made more pieces and pieces that go in different direction.
Pieces can now exit to the top or bottom in addition to the right.
Each direction has a straight hallway, an angled hallway, and a curve to
right.
Pieces with the entrance in the top have a room piece.
Created a start piece The just has an exit to the right
Made 3 end tiles for the directions the track pieces can exit to (no
pieces exit left). The exits are stairs, descending down.
Wall tiles are created higher than floor tiles now (hoped to stop the
player from getting out the side, but that isn't fixed).

In Player:
Upped maxMoveSpeed, lowered acceleration
Experimented with different movement types.
Some different styles make the car drift more (but too much).
Started trying to make the player bounce off of walls instead of just
coming to a dead stop.