Bug Fixes
Fixed a bug where enemies would spawn below the hallway with the top
entrance, bottom exit.

Added tileSpacing scaling to some enemy spawns
But not to skeletons, they work better bunched up

When hit by a projectile, the player get a 1 sec time penalty

Can now control the randomness with a seed

Tried scaling the map by 2 in all directions.
The mages in the hallways were less oppressive, but the map felt too big
and empty when driving the car. None of the corners were fun.

Increased the hitbox on the walls so they collide with the projectiles
(they were just flying right over top)