New Floor Tiles
Added a new floor tile and a feew variations
Made the map gen code work with random floor tiles

The floor tiles are more difficult to draw compared to the basic cubes.
Now I'm trying to find something to optimize to get back to a good fps.
The mages spawning a ton of fire balls seems to eb a main cause.

Made a large sphere that follows the player and turns mages on only when
inside the bubble.