Item Pool
Made an item pool for the mage projectiles
Instead of instantiating each ball and then deleting it, it just pulls one
from the item pool and gives it back when it's done with it for someone
else to use.

Also, enemies can now pass through one another.
They don't die when they are pushing into each other.
Surprisingly, this has less of an effect on fps than them all colliding
with each other and being destroyed as soon as you plow through the pile.