Checkered Lines
Added a checker pattern to start and finish lines.
Dragging a sprite onto a box puts the sprite on the side.
So When they get spawned, the lines need to be rotated onto their sides.
They also spawn 0.15 lower into the ground to hide the weird stretched
textures on the sides.