Menu
Added menu buttons to control the game
At the start, you can choose a specific seed or press a button to get a
random seed.
Then you can press Start to play the map.
After the map ends, the same buttons come up again (I think I want them to
look different in the final build).
The seed of the current map is loaded up.
Press Start to retry the same map
Or you can randomize the seed again.

To make this work, I had to cook up some spaghetti.
The map has a new empty object to store all the objects.
This storage object is the parent of every floor tile, wall, enemy, start,
and finish line. (instead of the map object being the parent as it was
before).
This object is deleted when the player presses the start button at the end
of the level.
This deletes the whole map.
It also deletes the player.
Then the map is built again from scratch.
The player is now Instantiated from the mapGen. The player used to just
start in the world.
Most of the spaghetti is getting references to everything the player
needs, adding them to MapGen, then giving them to the player when the
player is spawned. Also stuff that needs to know the player's position
like the camera and the MageWakeUp sphere.

Made the start and finish lines blend into the ground better. They also
have larger hit boxes (this was needed when the player had no gravity, but
the gravity is back).

Fixed some enemy facing issues. Some enemies would spawn and be facing the
wrong way (minor thing barely noticable). There are probably more that I
missed.