Seed in UI
Made an acorn sprite
The current seed is shown beside it on the screen.
Might like to make the number beside it the input box and click on the
acorn to randomize, but don't think that would be user friendly.
Unless there was text above them telling you that you could when the start
button is on screen.